see posts is working again

+++ public/js/delete.js
@@ -1,26 +1,31 @@
 // Delete your post
-document
-  .getElementById("delete-post")
-  .addEventListener("click", async (event) => {
-    event.preventDefault();
-    const userPostId = event.target.closest(".edit-post-form").dataset.postId;
+document.addEventListener("DOMContentLoaded", () => {
+  const deletePostButton = document.getElementById("delete-post");
 
-    if (confirm("Are you sure you want to delete this post?")) {
-      try {
-        const response = await fetch(`/api/post/delete/${userPostId}`, {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-          },
-        });
+  if (deletePostButton) {
+    deletePostButton.addEventListener("click", async (event) => {
+      event.preventDefault();
+      const userPostId = event.target.closest(".edit-post-form").dataset.postId;
 
-        if (response.ok) {
-          window.location.href = "/dashboard";
-        } else {
-          console.error("Error deleting post");
+      if (confirm("Are you sure you want to delete this post?")) {
+        try {
+          const response = await fetch(`/api/post/delete/${userPostId}`, {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+            },
+          });
+
+          if (response.ok) {
+            window.location.href = "/dashboard";
+          } else {
+            console.error("Error deleting post");
+          }
+        } catch (err) {
+          console.error(err);
         }
-      } catch (err) {
-        console.error(err);
       }
-    }
-  });
+    });
+  }
+});
+
